# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * مِرفق API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: string;
  isActive?: boolean;
  /** @nullable */
  unitId?: string | null;
  organizationId: string;
  organizationName?: string;
  brandColor?: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface MessageResponse {
  message: string;
}

export interface Property {
  id: string;
  name: string;
  address: string;
  managerId: string;
  createdAt: string;
  unitCount: number;
}

export interface PropertyInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  address: string;
}

export interface PropertyUpdate {
  name?: string;
  address?: string;
}

export interface Unit {
  id: string;
  number: string;
  floor: number;
  propertyId: string;
  /** @nullable */
  residentId?: string | null;
  /** @nullable */
  residentName?: string | null;
}

export interface UnitInput {
  /** @minLength 1 */
  number: string;
  floor: number;
  /** @nullable */
  residentId?: string | null;
}

export interface UnitUpdate {
  number?: string;
  floor?: number;
  /** @nullable */
  residentId?: string | null;
}

export interface MaintenanceRequest {
  id: string;
  title: string;
  description: string;
  category: string;
  priority: string;
  status: string;
  propertyId: string;
  /** @nullable */
  propertyName?: string | null;
  unitId: string;
  /** @nullable */
  unitNumber?: string | null;
  residentId: string;
  /** @nullable */
  residentName?: string | null;
  /** @nullable */
  technicianId?: string | null;
  /** @nullable */
  technicianName?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface RequestInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  category: string;
  priority: string;
  unitId: string;
}

export interface RequestUpdate {
  status?: string;
  priority?: string;
  /** @nullable */
  technicianId?: string | null;
  title?: string;
  description?: string;
}

export interface Comment {
  id: string;
  content: string;
  requestId: string;
  authorId: string;
  authorName: string;
  createdAt: string;
}

export interface CommentInput {
  /** @minLength 1 */
  content: string;
}

export interface TechnicianProfile {
  id: string;
  userId: string;
  name: string;
  specialty: string;
  phone: string;
  activeJobsCount: number;
}

export interface DashboardStats {
  totalRequests: number;
  pending: number;
  inProgress: number;
  completed: number;
  totalProperties: number;
  totalUnits: number;
  totalTechnicians: number;
}