# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * مِرفق API specification
 * OpenAPI spec version: 0.2.0
 */

export interface AuthUser {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: string;
  isActive?: boolean;
  /** @nullable */
  unitId?: string | null;
  organizationId: string;
  organizationName?: string;
  brandColor?: string;
}